requirements.txt
openai
langchain
tiktoken
python-dotenv
pandas
iPython
langchain[docarray]

